# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Plataforma de Compartilhamento de Receitas
  description: API para compartilhamento de receitas culinárias.
servers:
- url: http://localhost:3000
  description: Servidor local
...
paths:
...
  /recipes:
    get:
      tags:
      - Receitas
      summary: Obter todas as receitas
      description: Rota para obter todas as receitas cadastradas na plataforma.
      responses:
        "200":
          description: Receitas obtidas com sucesso!

    post:
      tags:
      - Receitas
      summary: Criar uma nova receita
      description: Rota para criar uma nova receita.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeCreate"
            examples:
              Example Json:
                value:
                  title: "Título da receita"
                  description: "Descrição da receita"
                  ingredients: ["Ingrediente 1", "Ingrediente 2"]
                  steps: ["Passo 1", "Passo 2"]
              Example Type:
                value:
                  title: "string"
                  description: "string"
                  ingredients: ["string"]
                  steps: ["string"]
      responses:
        "200":
          description: Receita criada com sucesso!
        "400":
          description: Erro ao criar receita.
  /recipes/{recipeId}:
    get:
      tags:
      - Receitas
      summary: Obter uma receita específica
      description: Rota para obter informações de uma receita específica.
      parameters:
      - name: recipeId
        in: path
        required: true
        description: ID da receita
        schema:
          type: integer
      responses:
        "200":
          description: Informações da receita obtidas com sucesso!
        "404":
          description: Receita não encontrada.

    put:
      tags:
      - Receitas
      summary: Atualizar uma receita
      description: Rota para atualizar uma receita existente.
      parameters:
      - name: recipeId
        in: path
        required: true
        description: ID da receita
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeUpdate"
            examples:
              Example Json:
                value:
                  title: "Novo título da receita"
                  description: "Nova descrição da receita"
                  ingredients: ["Ingrediente 1 atualizado", "Ingrediente 2 atualizado"]
                  steps: ["Passo 1 atualizado", "Passo 2 atualizado"]
              Example Type:
                value:
                  title: "string"
                  description: "string"
                  ingredients: ["string"]
                  steps: ["string"]
      responses:
        "200":
          description: Receita atualizada com sucesso!
        "400":
          description: Erro ao atualizar receita.
        "404":
          description: Receita não encontrada.

    delete:
      tags:
      - Receitas
      summary: Excluir uma receita
      description: Rota para excluir uma receita existente.
      parameters:
      - name: recipeId
        in: path
        required: true
        description: ID da receita
        schema:
          type: integer
      responses:
        "200":
          description: Receita excluída com sucesso!
        "404":
          description: Receita não encontrada.
...
components:
  schemas:
...
    RecipeCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
    RecipeUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
...